added trainer
