+++ pyproject.toml
@@ -13,6 +13,7 @@ datasets = "^2.5.1"
 black = "^22.8.0"
 dvc = "^2.27.2"
 dvc-s3 = "^2.20.1"
+fire = "^0.4.0"
 
 
 [build-system]
